% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-preprocessing.R
\name{add_areas.avoidance.multiple}
\alias{add_areas.avoidance.multiple}
\title{Adds areas of interest into the positioning data.}
\usage{
\method{add_areas}{avoidance.multiple}(obj, areas = default_zones())
}
\arguments{
\item{obj}{avoidance.multiple object}

\item{areas}{list of areas of interest to add. Defaults to implemented left, central and right areas}
}
\value{
object which was passed with added areas
}
\description{
Adds areas of interest into the positioning data.
}
