% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-const.R
\name{apparatus_image_path}
\alias{apparatus_image_path}
\title{Retuns path to the apparatus image}
\usage{
apparatus_image_path(darkside = RIGHT_ZONE_NAME)
}
\arguments{
\item{darkside}{which zone should be dark? Oprions are "left" or "right. 
Defaults to "right"}
}
\description{
Retuns path to the apparatus image
}
