% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-preprocessing.R
\name{as.navr}
\alias{as.navr}
\title{Converts the coordinate data to navr}
\usage{
as.navr(df)
}
\arguments{
\item{df}{data from load_table}
}
\value{
navr object
}
\description{
Converts the coordinate data to navr
}
