% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-analysis.R
\name{collect_freezes}
\alias{collect_freezes}
\alias{collect_freezes.avoidance.single}
\title{Searches for episodes of freezing behaviour}
\usage{
collect_freezes(obj, min_duration, speed_threshold, ...)

\method{collect_freezes}{avoidance.single}(obj, min_duration,
  speed_threshold, ...)
}
\arguments{
\item{obj}{valid object. Currently only avoidance single is recommended. The 
navr positioning data should optimally be smoothed for the best results to be
obtained. Use}

\item{min_duration}{minimal duration of the freeze}

\item{speed_threshold}{what movement speed is considered unmoving (there is still 
some movement in the machine)}

\item{...}{other optional parameters. see `navr::search_stops`` for the complete set}
}
\description{
This function is basically a complex wrapper around the navr::search_stops
function. The goal is to implement stable preprocessing nad smoothing here, so that 
it is not a mere wrapper around the navr::search_stops function
}
\section{Methods (by class)}{
\itemize{
\item \code{avoidance.single}: Implementation for avoidance.single object
}}

