% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-getters.R
\name{combine_all}
\alias{combine_all}
\title{Creates a single navr object with all animals combined together}
\usage{
combine_all(obj)
}
\arguments{
\item{obj}{avoidance.multiple object}
}
\value{

}
\description{
Creates a single navr object with all animals combined together
}
