% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-analysis.R
\name{create_crosses_vector}
\alias{create_crosses_vector}
\title{Creates a vector of length exp_length which contains number of crosses given area presence}
\usage{
create_crosses_vector(df_presence, exp_length = 3600)
}
\arguments{
\item{df_presence}{presence table calculated with \code{\link{get_area_presence}}}

\item{exp_length}{expected length of the experiment in seconds}
}
\value{
vector of length exp_length with number of crosses at that point
}
\description{
Creates a vector of length exp_length which contains number of crosses given area presence
}
