% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-visualisations.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Creates heatmap from passed object}
\usage{
create_heatmap(obj, bins = 100, background = apparatus_image_path(),
  ...)
}
\arguments{
\item{obj}{object to calculate heatmap from}

\item{bins}{number of bins}

\item{...}{optional params to the stat_density2d geom}

\item{geom}{which geom to use? possibile are "polygon"(default) or "raster"}
}
\description{
Creates heatmap from passed object
}
