% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-visualisations.R
\name{create_heatmap_polygon}
\alias{create_heatmap_polygon}
\title{Creates heatmap from passed navr object}
\usage{
create_heatmap_polygon(obj, bins = 100, ...)
}
\arguments{
\item{obj}{nav_object}
}
\description{
Creates heatmap from passed navr object
}
\examples{

}
