% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-getters.R
\name{filter_animals}
\alias{filter_animals}
\title{Filter the fgu.avoidance object for chosen animals}
\usage{
filter_animals(obj, codes)
}
\arguments{
\item{obj}{avoidance.multiple object}

\item{codes}{animal codes as they appear in the object}
}
\value{
filtered avoidance.multiple object or empty list
}
\description{
Filter the fgu.avoidance object for chosen animals
}
