% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-preprocessing.R
\name{filter_coordinates}
\alias{filter_coordinates}
\title{Filter out the coordinates from the loaded data}
\usage{
filter_coordinates(df)
}
\arguments{
\item{df}{}
}
\description{
Filter out the coordinates from the loaded data
}
