% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-analysis.R
\name{get_area_presence}
\alias{get_area_presence}
\title{Returns times of animal presence in each area of interest}
\usage{
get_area_presence(obj)
}
\arguments{
\item{obj}{avoidance.single object}
}
\value{
data.frame with times and animals location
}
\description{
Returns times of animal presence in each area of interest
}
