% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-getters.R
\name{get_position_table}
\alias{get_position_table}
\title{Returns positioning data}
\usage{
get_position_table(obj)
}
\arguments{
\item{obj}{avoidance.single object}
}
\value{

}
\description{
Returns positioning data
}
