% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-loading.R
\name{load_data}
\alias{load_data}
\title{Load data from a particular filepath}
\usage{
load_data(filepath = NULL, text = NULL)
}
\arguments{
\item{filepath}{Either filepath or a text}

\item{text}{alternatively the data can be loaded from loaded text}
}
\value{
avoidance.single object
}
\description{
The data are loaded and then preprocessed. 
Areas are added (see \code{\link{add_areas}}), position data are converted to 
navr object and classes are added to the resulting objects
}
