% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-loading.R
\name{load_folder}
\alias{load_folder}
\title{Loads all files from a folder and combines to a large table which is then processed}
\usage{
load_folder(folder)
}
\arguments{
\item{folder}{}
}
\description{
Loads all files from a folder and combines to a large table which is then processed
}
