% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-loading.R
\name{load_text_data}
\alias{load_text_data}
\title{Loads text data from already loaded text}
\usage{
load_text_data(text)
}
\arguments{
\item{text}{}
}
\description{
Loads text data from already loaded text
}
