% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-visualisations.R
\name{plot_area_presence}
\alias{plot_area_presence}
\title{PLots an image of area presence for avoidance.single}
\usage{
plot_area_presence(obj)
}
\arguments{
\item{obj}{avoidance.single object}
}
\description{
PLots an image of area presence for avoidance.single
}
