% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-visualisations.R
\name{plot_path.avoidance.single}
\alias{plot_path.avoidance.single}
\title{Creates a path graph of a single trial}
\usage{
\method{plot_path}{avoidance.single}(obj, center = central_zone(),
  background = NULL, color = "#98959a", size = 1.25, ...)
}
\arguments{
\item{obj}{avoidance single object}

\item{center}{}

\item{background}{}

\item{color}{}

\item{size}{}

\item{...}{}
}
\description{
Creates a path graph of a single trial
}
