% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-visualisations.R
\name{plot_speed.avoidance.single}
\alias{plot_speed.avoidance.single}
\title{Plots animal's speed}
\usage{
\method{plot_speed}{avoidance.single}(obj, scaling = "none",
  constraints = NULL, ...)
}
\arguments{
\item{obj}{avoidnace.single object}

\item{scaling}{scaling to implement. see `navr::plot_speed` for specifics}

\item{constraints}{if the speed is scaled, it can be constrained to defined values. 
Requires numeric(2)}

\item{...}{other \code{\link{ggplot::geom_line}} parameters}
}
\value{
ggplot plot object
}
\description{
Plots animal's speed
}
