% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-analysis.R
\name{session_results}
\alias{session_results}
\title{Basic results to give a report from a single session}
\usage{
session_results(obj)
}
\arguments{
\item{obj}{either avoidance.single or avoidance.multiple object}
}
\value{
data frame with basic session results
}
\description{
Basic results to give a report from a single session
}
