% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-preprocessing.R
\name{smooth_positions.avoidance.single}
\alias{smooth_positions.avoidance.single}
\title{Smooths positioning data}
\usage{
\method{smooth_positions}{avoidance.single}(obj, ...)
}
\arguments{
\item{obj}{avoidance.single object}

\item{...}{parameters for the `navr::smooth_positions` function.}
}
\description{
S3 method for the `navr::smooth_positions`.`
}
