% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-preprocessing.R
\name{smooth_speed.avoidance.single}
\alias{smooth_speed.avoidance.single}
\title{Smooths speeds}
\usage{
\method{smooth_speed}{avoidance.single}(obj, ...)
}
\arguments{
\item{obj}{avoidance.single object}

\item{...}{parameters for the `navr::smooth_speed`` function}
}
\description{
S3 method for the `navr::smooth_speed`
}
