% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-preprocessing.R
\name{add_areas.avoidance.single}
\alias{add_areas.avoidance.single}
\title{Adds areas into the positioning data}
\usage{
\method{add_areas}{avoidance.single}(obj, areas = default_zones())
}
\arguments{
\item{obj}{avoidance.single object}

\item{areas}{list of areas of interest to add. *Defaults* to implemented left, central and right areas 
in the function \code{\link{default_zones}}}
}
\value{
object which was passed with added areas
}
\description{
Adds areas into the positioning data
}
