% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-analysis.R
\name{collect_crosses}
\alias{collect_crosses}
\title{Collects information about each cross in given object}
\usage{
collect_crosses(obj)
}
\value{
dataframe with results
}
\description{
Collects information about each cross in given object
}
