% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-analysis.R
\name{collect_crosses}
\alias{collect_crosses}
\alias{collect_crosses.avoidance.multiple}
\alias{collect_crosses.avoidance.single}
\title{Collects information about each cross in given object}
\usage{
collect_crosses(obj)

\method{collect_crosses}{avoidance.multiple}(obj)

\method{collect_crosses}{avoidance.single}(obj)
}
\arguments{
\item{obj}{avoidance single or multiple object with areas added.}
}
\value{
dataframe with *from, to, time, index* columns. From defines which area was the cross from, to
which area it was made to. Time is the time of the cross and index is the index in the position data
}
\description{
The function calculates crosses from each of the default areas 
definined in the object except the central area. The central area is considered 
a "buffer" - if an animal enters only the central area, cross is NOT counted. 
The areas need to be added during the preprocessing and area generally automatically 
added with \code{\link{load_data}}. If you need to define your own areas you
can do so using the \code{\link{add_areas}} function.

In case you want to use different then the default areas, you either need 
to define your own or just use the `get_area_visits` from the `navr` package, 
which this function uses under the hood
}
\section{Methods (by class)}{
\itemize{
\item \code{avoidance.multiple}: calculates all the crosses for all the animals in the 
object.

\item \code{avoidance.single}: calculates the crosses for the animal. Wrapper around
the \code{\link{navr::get_area_visits}}.
}}

