% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-analysis.R
\name{collect_crosses}
\alias{collect_crosses}
\title{Collects information about each cross in given object}
\usage{
collect_crosses(obj)
}
\value{
dataframe with *from, to, time, index* columns. From defines whihc area was the cross from, to
which area it was made to. Time is the time of the cross and index is the index in the position data
}
\description{
Uses `get_area_visits` form the package `navr` under the hood.
}
