% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-visualisations.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Creates heatmap from passed object}
\usage{
create_heatmap(obj, bins = 100, geom = "polygon", ...)
}
\arguments{
\item{obj}{object to calculate heatmap from}

\item{bins}{number of bins}

\item{geom}{which geom to use? possibile are "polygon"(default) or "raster"}

\item{...}{optional params to the stat_density2d geom}
}
\description{
Creates heatmap from passed object
}
