% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-loading.R
\name{load_data}
\alias{load_data}
\title{Load data from a particular filepath}
\usage{
load_data(filepath = NULL, text = NULL)
}
\arguments{
\item{filepath}{Either filepath or a text}

\item{text}{althernatively the data can be loaded from loaded text}
}
\value{
avoidance.single object
}
\description{
Load data from a particular filepath
}
