% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-loading.R
\name{load_data}
\alias{load_data}
\title{Load data from a particular filepath}
\usage{
load_data(filepath)
}
\arguments{
\item{filepath}{path to the data.csv}
}
\value{
avoidance.single object
}
\description{
Load data from a particular filepath
}
