% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-visualisations.R
\name{plot_path.avoidance.single}
\alias{plot_path.avoidance.single}
\title{Creates a path graph of a single trial}
\usage{
\method{plot_path}{avoidance.single}(obj, zone = central_zone())
}
\arguments{
\item{obj}{}

\item{...}{}
}
\description{
Creates a path graph of a single trial
}
