% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoidance-visualisations.R
\name{plot_path.avoidance.single}
\alias{plot_path.avoidance.single}
\title{Creates a path graph of a single trial}
\usage{
\method{plot_path}{avoidance.single}(obj,
  background = apparatus_image_path(), center = FALSE,
  color = "#98959a", size = 1.25, ...)
}
\arguments{
\item{obj}{avoidance single object}

\item{background}{path to a backgroud image to be plotted. Use \code{\link{apparatus_image_path}} 
to construct or provide your own.}

\item{center}{logical determining if the central zone should be plotted}

\item{color}{color of the path}

\item{size}{path size}

\item{...}{}
}
\description{
Creates a path graph of a single trial
}
